added shortcuts in vscode.md (more info in deskcription)
added general and basic editing shortcuts for vs code. I added those in the Linux category, I am not sure if they are the same for other OSs but in my resources, they said they are for Linux so if there are mistakes please correct them. Also, let me know if I should add more shortcuts.